+++ pytouch/boilerplate/__PROJECT_NAME__/__PROJECT_NAME__.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-# This is the core of __PROJECT_NAME__
-
-# Author: Your Name <your@name.me>
-# (c) 2022 Some rights reversed.
-
-