+++ pytouch/boilerplate/__PROJECT_NAME__/__init__.py
@@ -1,9 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-# This is the __init__.py file of __PROJECT_NAME__
-# If it does anything, it is important to document 
-# what and why
-
-# Author: Your Name <your@name.me>
-# (c) 2022 Some rights reversed.