+++ pytouch/boilerplate/__PROJECT_NAME__/framework/__init__.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-# This is the __init__.py file of subdirectory and thereby module /framework
-# of the __PROJECT_NAME__ 
-
-# Author: Your Name <your@name.me>
-# (c) 2022 Some rights reversed.