+++ CHANGES.md
@@ -1,6 +1,12 @@
 CHANGES
 =======
 
+[07/10/2022]
+* increase documentation
+* add more than one template into /static_templates
+* make a small utility the standard, not the framework
+
+
 [03/10/2022]
 * Created a CHANGES.md
 * Rename of cfg.yaml to .cfg.yaml